backend/requirements.txt
fastapi[standard]
langchain
langchain-openai
openai
python-dotenv
langchain-community